requirements.txt
# Requirements for ProxyMaster 2.0
# Core packages - Required for basic functionality
httpx>=0.23.0
beautifulsoup4>=4.11.1
PySocks>=1.7.1
rich>=12.0.0
requests>=2.27.1
schedule>=1.1.0
aiohttp>=3.8.1
asyncio>=3.4.3
pandas>=1.3.5

# GUI packages - Required for graphical interface
ttkthemes>=3.2.0
customtkinter>=4.6.0
pillow>=9.0.0

# Data visualization - For analytics features
matplotlib>=3.5.1
seaborn>=0.11.2

# Geolocation - For proxy geolocation features
geoip2>=4.6.0
maxminddb>=2.2.0

# Database - For analytics storage
sqlite3-api>=0.1.0
matplotlib>=3.5.1

# GUI packages (optional)
customtkinter>=5.2.0
ttkthemes>=3.2.0

# Utility packages
colorama>=0.4.4
urllib3>=1.26.9
lxml>=4.9.0

# Development packages (optional)
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.1